+++ source/constructs/lib/serverless-image-handler.ts
@@ -81,6 +81,21 @@ export class ServerlessImageHandler extends Construct {
       });
       enableDefaultFallbackImageCondition.overrideLogicalId('EnableDefaultFallbackImageCondition');
 
+      const isOptInRegion = new cdk.CfnCondition(this, 'IsOptInRegion', {
+        expression: cdk.Fn.conditionOr(
+          cdk.Fn.conditionEquals("af-south-1", cdk.Aws.REGION),
+          cdk.Fn.conditionEquals("ap-east-1", cdk.Aws.REGION),
+          cdk.Fn.conditionEquals("eu-south-1" , cdk.Aws.REGION),
+          cdk.Fn.conditionEquals("me-south-1" , cdk.Aws.REGION)
+        )
+      });
+      isOptInRegion.overrideLogicalId('IsOptInRegion');
+      
+      const isNotOptInRegion = new cdk.CfnCondition(this, 'IsNotOptInRegion', {
+        expression: cdk.Fn.conditionNot(isOptInRegion)
+      });
+      isNotOptInRegion.overrideLogicalId('IsNotOptInRegion')
+
       // ImageHandlerFunctionRole
       const imageHandlerFunctionRole = new cdkIam.Role(this, 'ImageHandlerFunctionRole', {
         assumedBy: new cdkIam.ServicePrincipal('lambda.amazonaws.com'),
@@ -122,7 +137,8 @@ export class ServerlessImageHandler extends Construct {
           }),
           new cdkIam.PolicyStatement({
             actions: [
-              'rekognition:DetectFaces'
+              'rekognition:DetectFaces',
+              'rekognition:DetectModerationLabels'
             ],
             resources: [
               '*'
@@ -182,6 +198,12 @@ export class ServerlessImageHandler extends Construct {
       });
       const cfnLambdaFunctionLogs = lambdaFunctionLogs.node.defaultChild as cdkLogs.CfnLogGroup;
       cfnLambdaFunctionLogs.retentionInDays = props.logRetentionPeriodParameter.valueAsNumber;
+      this.addCfnNagSuppressRules(cfnLambdaFunctionLogs, [
+        {
+          "id": "W84",
+          "reason": "Used to store store function info"
+        }
+      ]);
       cfnLambdaFunctionLogs.overrideLogicalId('ImageHandlerLogGroup');
 
       // CloudFrontToApiGatewayToLambda pattern
@@ -193,6 +215,16 @@ export class ServerlessImageHandler extends Construct {
 
       // ApiLogs
       const cfnApiGatewayLogGroup = apiGatewayLogGroup.node.defaultChild as cdkLogs.CfnLogGroup;
+      this.addCfnNagSuppressRules(cfnApiGatewayLogGroup, [
+        {
+          "id": "W84",
+          "reason": "Used to store store api log info, not using kms"
+        },
+        {
+          "id": "W86",
+          "reason": "Log retention specified in CloudFromation parameters."
+        }
+      ]);
       cfnApiGatewayLogGroup.overrideLogicalId('ApiLogs');
 
       // ImageHandlerApi
@@ -242,6 +274,7 @@ export class ServerlessImageHandler extends Construct {
       const cloudFrontToApiGateway = cloudFrontApiGatewayLambda.node.findChild('CloudFrontToApiGateway');
       const accessLogBucket = cloudFrontToApiGateway.node.findChild('CloudfrontLoggingBucket') as cdkS3.Bucket;
       const cfnAccessLogBucket = accessLogBucket.node.defaultChild as cdkS3.CfnBucket;
+      cfnAccessLogBucket.cfnOptions.condition = isNotOptInRegion;
       this.addCfnNagSuppressRules(cfnAccessLogBucket, [
         {
           "id": "W35",
@@ -252,8 +285,76 @@ export class ServerlessImageHandler extends Construct {
 
       // LogsBucketPolicy
       const accessLogBucketPolicy = accessLogBucket.node.findChild('Policy') as cdkS3.BucketPolicy;
+      const cfnAccessLogBucketPolicy = accessLogBucketPolicy.node.defaultChild as cdkS3.CfnBucketPolicy;
+      (accessLogBucketPolicy.node.defaultChild as cdkS3.CfnBucketPolicy).cfnOptions.condition = isNotOptInRegion;
       (accessLogBucketPolicy.node.defaultChild as cdkS3.CfnBucketPolicy).overrideLogicalId('LogsBucketPolicy');
 
+      //OptInRegionLogBucket
+      const optInRegionAccessLogBucket = cdkS3.Bucket.fromBucketAttributes(this, 'CloudFrontLoggingBucket', {
+        bucketName: 
+          cdk.Fn.getAtt(
+            cdk.Lazy.stringValue({ 
+              produce(context) {
+                return cfLoggingBucket.logicalId}
+            }), 
+          'bucketName').toString(),
+         region: 'us-east-1'
+      });
+      
+      //OptInRegionLogBucketPolicy
+      const optInRegionPolicyStatement = cfnAccessLogBucketPolicy.policyDocument.toJSON().Statement[0];
+      optInRegionPolicyStatement.Resource = "";
+
+      //Choose Log Bucket
+      const cloudFrontLogsBucket = cdk.Fn.conditionIf(isOptInRegion.logicalId, optInRegionAccessLogBucket.bucketRegionalDomainName, accessLogBucket.bucketRegionalDomainName).toString();
+      
+      
+      //ImagehandlerCachePolicy
+      const cfnCachePolicy = new cdkCloudFront.CfnCachePolicy(
+        this, 
+        'CachePolicy', 
+        {
+          cachePolicyConfig: {
+            name: `${cdk.Aws.STACK_NAME}-${cdk.Aws.REGION}-ImageHandlerCachePolicy`,
+            defaultTtl: 86400,
+            minTtl: 1,
+            maxTtl: 31536000,
+            parametersInCacheKeyAndForwardedToOrigin: {
+              cookiesConfig: {cookieBehavior: "none"},
+              enableAcceptEncodingGzip: true,
+              headersConfig: {
+                headerBehavior: "whitelist", 
+                headers:['origin', 'accept']
+              },
+              queryStringsConfig: {
+                queryStringBehavior: "whitelist",
+                queryStrings: ["signature"]
+              },
+            }       
+        }
+      });
+      cfnCachePolicy.overrideLogicalId("ImageHandlerCachePolicy");
+
+      //ImageHandlerOriginRequestPolicy
+      const cfnOriginRequestPolicy = new cdkCloudFront.CfnOriginRequestPolicy(
+        this, 
+        "OriginRequestPolicy",
+        {
+          originRequestPolicyConfig: { 
+            cookiesConfig: {cookieBehavior: "none"},
+            headersConfig: {
+              headerBehavior: "whitelist",
+              headers: ['origin', 'accept']
+            },
+            name: `${cdk.Aws.STACK_NAME}-${cdk.Aws.REGION}-ImageHandlerOriginRequestPolicy`,
+            queryStringsConfig: {
+              queryStringBehavior: "whitelist",
+              queryStrings: ["signature"]
+            },
+          }
+        });
+        cfnOriginRequestPolicy.overrideLogicalId("ImageHandlerOriginRequestPolicy");
+ 
       // ImageHandlerDistribution
       const cfnCloudFrontDistribution = cloudFrontWebDistribution.node.defaultChild as cdkCloudFront.CfnDistribution;
       cfnCloudFrontDistribution.distributionConfig = {
@@ -273,13 +374,10 @@ export class ServerlessImageHandler extends Construct {
         defaultCacheBehavior: {
           allowedMethods: [ 'GET', 'HEAD' ],
           targetOriginId: apiGateway.restApiId,
-          forwardedValues: {
-            queryString: true,
-            queryStringCacheKeys: [ 'signature' ],
-            headers: [ 'Origin', 'Accept' ],
-            cookies: { forward: 'none' }
-          },
-          viewerProtocolPolicy: 'https-only'
+          viewerProtocolPolicy: 'https-only',
+          cachePolicyId: cfnCachePolicy.ref,
+          originRequestPolicyId: cfnOriginRequestPolicy.ref
+
         },
         customErrorResponses: [
           { errorCode: 500, errorCachingMinTtl: 10 },
@@ -291,7 +389,7 @@ export class ServerlessImageHandler extends Construct {
         priceClass: 'PriceClass_All',
         logging: {
           includeCookies: false,
-          bucket: accessLogBucket.bucketRegionalDomainName,
+          bucket: cloudFrontLogsBucket,
           prefix: 'image-handler-cf-logs/'
         }
       };
@@ -378,7 +476,7 @@ export class ServerlessImageHandler extends Construct {
         httpVersion: 'http2',
         logging: {
           includeCookies: false,
-          bucket: accessLogBucket.bucketRegionalDomainName,
+          bucket: cloudFrontLogsBucket,
           prefix: 'demo-cf-logs/'
         }
       };
@@ -416,6 +514,10 @@ export class ServerlessImageHandler extends Construct {
           }),
           new cdkIam.PolicyStatement({
             actions: [
+              's3:putBucketAcl',
+              's3:putEncryptionConfiguration',
+              's3:putBucketPolicy',
+              's3:CreateBucket',
               's3:GetObject',
               's3:PutObject',
               's3:ListBucket'
@@ -461,6 +563,12 @@ export class ServerlessImageHandler extends Construct {
       });
       const cfnCustomResourceLogGroup = customResourceLogGroup.node.defaultChild as cdkLogs.CfnLogGroup;
       cfnCustomResourceLogGroup.retentionInDays = props.logRetentionPeriodParameter.valueAsNumber;
+      this.addCfnNagSuppressRules(cfnCustomResourceLogGroup, [
+        {
+          "id": "W84",
+          "reason": "Used to store store function info, no kms used"
+        }
+      ]);
       cfnCustomResourceLogGroup.overrideLogicalId('CustomResourceLogGroup');
 
       // CustomResourceCopyS3
@@ -564,6 +672,19 @@ export class ServerlessImageHandler extends Construct {
         ],
         condition: enableDefaultFallbackImageCondition,
         dependencies: [ cfnCustomResourceRole, cfnCustomResourcePolicy ]
+      });
+
+      const bucketSuffix = cdk.Aws.STACK_NAME + cdk.Aws.REGION + cdk.Aws.ACCOUNT_ID;
+      const cfLoggingBucket = this.createCustomResource('CustomCFLoggingBucket', customResourceFunction, {
+        properties: [
+         { path: 'customAction', value: 'createCFLoggingBucket' },
+         { path: 'stackName', value: cdk.Aws.STACK_NAME },
+         { path: 'bucketSuffix', value: bucketSuffix },
+         { path: 'policy', value: optInRegionPolicyStatement }
+        ],
+        condition: isOptInRegion,
+        dependencies: [ cfnCustomResourceRole, cfnCustomResourcePolicy ]
+
       });
     } catch (error) {
       console.error(error);